Fix backfill filtering logic and marker format
It adds spaces around the asterisks of created markers to distinguish them from Whisper's markers.

Makes sure that segments are only filtered when they fully match one of the ignore words.

The filtering logic is updated

Enhances the disfluency marker creation logic to merge consecutive filtered words into a single marker (may not be necessary in the end)